requirements.txt
pyexcel-xls==0.2.3
googlefinance==0.7
google-api-python-client==1.5.5
redis==2.10.5
python-twitter==3.2
-e git+git@github.com:charlierguo/gmail.git#egg=gmail
